+++ Headers/VideoModel/VideoModel.h
@@ -3,9 +3,9 @@
 #include "Core/Core.h"
 
 struct VideoModel {
-    int SegmentDurationInMs;
-    std::vector<int> BitRatesInKbps;
-    std::vector<std::vector<int>> SegmentByteCounts;
+    size_t SegmentDurationInMs;
+    std::vector<size_t> BitRatesInKbps;
+    std::vector<std::vector<size_t>> SegmentByteCounts;
 
     static VideoModel Import(std::string_view fileName);
 };
